+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1 (2019-06-??)
 - add published lock file: prevents tagging/dist/upload after publish
+- tag before sdist if no tag
 
 ## 0.0 (2019-06-08)
 * Initially created.