+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1 (2019-06-??)
 - add published lock file: prevents tagging/dist/upload after publish
 - tag before sdist if no tag
+- push tag after publishing
 
 ## 0.0 (2019-06-08)
 * Initially created.